Same fetching logic used for `getStaticPaths`
