+++ pages/index.tsx
@@ -3,41 +3,12 @@ import Head from "next/head";
 import { Navbar } from "../components/navbar";
 import VideoCard from "../components/videoCard";
 import { IVideo } from "../types";
+import { fetchVideos } from "../utlis/fetchVideos";
 
 export async function getServerSideProps() {
-  const apiKey = process.env.NEXT_PUBLIC_YOUTUBE_API;
-  const searchQuery = encodeURIComponent("coding, podcast, upsc");
-  const maxResults = 100;
-  const regionCode = "in";
+  const videos = await fetchVideos();
 
-  const url = `https://youtube.googleapis.com/youtube/v3/search?part=snippet&maxResults=${maxResults}&q=${searchQuery}&key=${apiKey}&regionCode=${regionCode}`;
-
-  try {
-    const response = await fetch(url);
-
-    if (!response.ok) {
-      throw new Error(`HTTP error! Status: ${response.status}`);
-    }
-
-    const data = await response.json();
-
-    const videos = data.items
-      .filter((video: any) => video.id.videoId)
-      .map((video: any) => ({
-        videoId: video.id.videoId,
-        channelId: video.snippet.channelId,
-        channelTitle: video.snippet.channelTitle,
-        description: video.snippet.description,
-        publishTime: video.snippet.publishTime,
-        thumbnail: video.snippet.thumbnails.medium.url,
-        title: video.snippet.title,
-      }));
-
-    return { props: { videos } };
-  } catch (error) {
-    console.error("Fetch error:", error);
-    return { props: { videoData: [] } };
-  }
+  return { props: { videos } };
 }
 
 export default function Page({ videos }: { videos: IVideo[] }) {
@@ -56,7 +27,7 @@ export default function Page({ videos }: { videos: IVideo[] }) {
         minChildWidth="17rem"
         spacing="8"
       >
-        {videos.map((video) => {
+        {videos.map((video: IVideo) => {
           return (
             video.videoId && (
               <VideoCard
